Add ability to abort from Vim exit code
In many cases where Vim is used as an intermediate step in a series of
steps it is helpful to abort when the user explicitly exits vim with a
non-zero exit code.

What that means in this project is that if the user who is editing the
buffer decides that he/she does NOT wish to overwrite the clipboard
(maybe they realized they have something important) they would quit Vim
with `:cq` instead of the normal ways. This means they can write the
file all they want and it will not be copied to the clipboard if they
use the `:cq` command.

Granted the utility of this is debatable making this change optional.