Prune temp files manually
Add a utility to prune the temp files manually. By default this would be
history older then seven days but can be adjusted through a command line
option.

Originally intended to be used in a cron job for systems that either
don't reboot often or are not configured to erase the `/tmp` folder on
reboot.