Support capital 'Darwin' $OSTYPE
On some macOS environments, the `$OSTYPE` environment variable is
'Darwin' (capitalized) and not 'darwin'. This caused the install script
to echo "OS Darwin is not supported!" and exit, even if everything was
in fact in order.

This patch uses a case-insensitive regex against a [standard bash
operator](1) to allow any `$OSTYPE` that starts with 'Darwin' or
'darwin'.

All other instances of conditional checks for darwin* are also patched
across the code base.

[1]: https://www.gnu.org/software/bash/manual/bash.html#index-_005d_005d